admin features + data validation
Admin features: adding new restaurants and new food menus through UI.
Data validation: where the data should be a number, is verified if it is so. The name has to be longer than 3 characters.

To do: after adding a new restaurant, to display it in the drop menu options, the page needs to be refreshed(not a big deal :D )